moved some generic parts of file_io functions to io_utils; correction in _add_motor_impedances_ppc
