+++ pandapower/build_bus.py
@@ -561,7 +561,8 @@ def _add_motor_impedances_ppc(net, ppc):
     z_motor = 1 / (motor.sn_mva.values * 1e-3) / motor.k  # vn_kv**2 becomes 1**2=1 in per unit
     x_motor = z_motor / np.sqrt(motor.rx ** 2 + 1)
     r_motor = motor.rx * x_motor
-    y_motor = 1 / (r_motor + x_motor * 1j)
+    r_motor_np, x_motor_np = r_motor.values, x_motor.values
+    y_motor = 1 / (r_motor_np + x_motor_np * 1j)
 
     buses, gs, bs = _sum_by_group(motor_buses_ppc, y_motor.real, y_motor.imag)
     ppc["bus"][buses, GS] = gs
